+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 ### Changed
 - increased small button height
 - highlicht "smartphones" on desktop LP instead of URL
+- changed to heart BVG logo
+- updated desktop LP text and images
+- changed on scroll animations (desktop LP) to automatic animations for smaller screens
 
 ## [0.18.0] - 2020-07-29
 ### Added